small reorg of lighteval task class to allow mocking
