+++ src/lighteval/tasks/lighteval_task.py
@@ -26,7 +26,7 @@ import random
 from dataclasses import asdict, dataclass, field
 from typing import Callable
 
-from datasets import DatasetDict
+from datasets import DatasetDict, load_dataset
 from huggingface_hub import TextGenerationInputGrammarType
 from multiprocess import Pool
 from pytablewriter import MarkdownTableWriter
@@ -36,7 +36,7 @@ from lighteval.tasks.prompt_manager import FewShotSampler
 from lighteval.tasks.requests import (
     Doc,
 )
-from lighteval.utils.utils import ListLike, as_list, download_dataset_worker
+from lighteval.utils.utils import ListLike, as_list
 
 
 logger = logging.getLogger(__name__)
@@ -241,7 +241,7 @@ class LightevalTask:
             list[Doc]: List of documents.
         """
         if self.dataset is None:
-            self.dataset = download_dataset_worker(
+            self.dataset = self.download_dataset_worker(
                 self.dataset_path,
                 self.dataset_config_name,
                 self.trust_dataset,
@@ -356,34 +356,42 @@ class LightevalTask:
         """
 
         if dataset_loading_processes <= 1:
-            datasets = [
-                download_dataset_worker(
-                    task.dataset_path,
-                    task.dataset_config_name,
-                    task.trust_dataset,
-                    task.dataset_filter,
-                    task.dataset_revision,
-                )
-                for task in tasks.values()
-            ]
+            # Useful for the test suite: we can mock loading tasks by overwriting the
+            # individual download_dataset_worker functions
+            datasets = [task.download_dataset_worker(task) for task in tasks.values()]
         else:
             with Pool(processes=dataset_loading_processes) as pool:
                 datasets = pool.starmap(
-                    download_dataset_worker,
-                    [
-                        (
-                            task.dataset_path,
-                            task.dataset_config_name,
-                            task.trust_dataset,
-                            task.dataset_filter,
-                            task.dataset_revision,
-                        )
-                        for task in tasks.values()
-                    ],
+                    LightevalTask.download_dataset_worker,
+                    [tasks.values()],
                 )
 
         for task, dataset in zip(tasks, datasets):
             tasks[task].dataset = dataset
+
+    @staticmethod
+    def download_dataset_worker(
+        task: "LightevalTask",
+    ) -> DatasetDict:
+        """
+        Worker function to download a dataset from the HuggingFace Hub.
+        Used for parallel dataset loading.
+        """
+        dataset = load_dataset(
+            path=task.dataset_path,
+            name=task.dataset_config_name,
+            data_dir=None,
+            cache_dir=None,
+            download_mode=None,
+            trust_remote_code=task.trust_dataset,
+            revision=task.dataset_revision,
+        )
+
+        if task.dataset_filter is not None:
+            dataset = dataset.filter(task.dataset_filter)
+
+        # It returns DatasetDict because we don't specify a split
+        return dataset  # type: ignore
 
 
 def extract_num_samples(metric_name: str) -> int:
